bug(rearrange-project-structure): fix create new file called validation put all validation function in validation file put all helper in their respective files remove the unecessary files from repo wrap article helper function inside the classes wrap user helper function inside the classes wrap controller inside a class [Finishes #168899159]
