--- .nyc_output/processinfo/01dc0b9c-2987-46f1-a1ca-c7e53a1078f7.json
@@ -1,1 +1,0 @@
-{"uuid":"01dc0b9c-2987-46f1-a1ca-c7e53a1078f7","parent":null,"pid":1816,"argv":["C:\\Program Files\\nodejs\\node.exe","D:\\SETH\\Teamwork\\node_modules\\mocha\\bin\\mocha","--require","@babel/register","server/tests/*.spec.js","--exit"],"execArgv":[],"cwd":"D:\\SETH\\Teamwork","time":1569784822685,"ppid":2700,"root":"fc4e8b3e-14f9-4b53-969b-b5c1197b3cdb","coverageFilename":"D:\\SETH\\Teamwork\\.nyc_output\\01dc0b9c-2987-46f1-a1ca-c7e53a1078f7.json","files":["D:\\SETH\\Teamwork\\server\\tests\\articles.spec.js","D:\\SETH\\Teamwork\\server\\app.js","D:\\SETH\\Teamwork\\server\\routes\\route.js","D:\\SETH\\Teamwork\\server\\controllers\\userController.js","D:\\SETH\\Teamwork\\server\\helpers\\userHelper.js","D:\\SETH\\Teamwork\\server\\models\\user.js","D:\\SETH\\Teamwork\\server\\models\\article.js","D:\\SETH\\Teamwork\\server\\controllers\\articleController.js","D:\\SETH\\Teamwork\\server\\helpers\\articleHelper.js","D:\\SETH\\Teamwork\\server\\middlewares\\authentication.js","D:\\SETH\\Teamwork\\server\\tests\\sginup.spec.js"]}
