--- .nyc_output/processinfo/index.json
@@ -1,1 +1,0 @@
-{"processes":{"01dc0b9c-2987-46f1-a1ca-c7e53a1078f7":{"parent":null,"children":[]}},"files":{"D:\\SETH\\Teamwork\\server\\tests\\articles.spec.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\app.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\routes\\route.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\controllers\\userController.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\helpers\\userHelper.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\models\\user.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\models\\article.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\controllers\\articleController.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\helpers\\articleHelper.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\middlewares\\authentication.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"],"D:\\SETH\\Teamwork\\server\\tests\\sginup.spec.js":["01dc0b9c-2987-46f1-a1ca-c7e53a1078f7"]},"externalIds":{}}
