+++ server/controllers/articleController.js
@@ -2,11 +2,12 @@ import moment from 'moment';
 import uuidv1 from 'uuidv1';
 import articles from '../models/article';
 import ArticleHelper from '../helpers/articleHelper';
+import Validation from '../helpers/validation';
 import helper from '../helpers/userHelper';
 
 const articleController = {
   shareArticles(req, res) {
-    const { error } = helper.validateArticle(req.body);
+    const { error } = Validation.validateArticle(req.body);
     if (error) {
       res.status(400).json({
         status: 400,
@@ -15,20 +16,20 @@ const articleController = {
     }
 
     // eslint-disable-next-line max-len
-    const article = new ArticleHelper(req.body.authorId, req.body.authorName, req.body.title, req.body.article);
-    articles.push(article);
+    const article = req.body;
+    ArticleHelper.AddArticle(article);
     res.status(200).json({
       status: 200,
       message: 'article successfully created',
       data: {
         createdOn: moment().format('YY-MM-DD H:m'),
         title: article.title,
-        articleId: article.articleId,
+        articleId: articles[articles.length - 1].articleId,
       },
     });
   },
   viewOneArticle(req, res) {
-    const article = helper.findArticle(req.params.id);
+    const article = ArticleHelper.findArticle(req.params.id);
     if (!article) {
       return res.status(404).json({
         status: 404,
@@ -50,7 +51,7 @@ const articleController = {
     });
   },
   editArticle(req, res) {
-    const articleToEdit = helper.findArticle(req.params.id);
+    const articleToEdit = ArticleHelper.findArticle(req.params.id);
 
     if (!articleToEdit) {
       return res.status(404).json({
@@ -58,7 +59,7 @@ const articleController = {
         message: 'article with given id is not found',
       });
     }
-    if (!helper.checkAuthor(req)) {
+    if (!ArticleHelper.checkAuthor(req)) {
       return res.status(401).json({
         status: 401,
         message: 'user is not allowed to edit this article',
@@ -82,9 +83,9 @@ const articleController = {
   deleteArticle(req, res) {
     let status;
     let message;
-    const toDelete = helper.findArticle(req.params.id);
+    const toDelete = ArticleHelper.findArticle(req.params.id);
     if (toDelete) {
-      if (helper.checkAuthor(req)) {
+      if (ArticleHelper.checkAuthor(req)) {
         const index = articles.indexOf(toDelete);
         articles.splice(index, 1);
         return res.sendStatus(204);
@@ -103,7 +104,7 @@ const articleController = {
   addComment(req, res) {
     let status;
     let message;
-    const { error } = helper.validateComment(req.body);
+    const { error } = Validation.validateComment(req.body);
     if (error) {
       status = 400;
       message = error.details[0].message.replace(/[/"]/g, '');
@@ -112,7 +113,7 @@ const articleController = {
         message,
       });
     }
-    const articleToComment = helper.findArticle(req.params.id);
+    const articleToComment = ArticleHelper.findArticle(req.params.id);
     if (articleToComment) {
       const author = helper.getUser(req.payload.email);
       const userComment = {
