+++ server/helpers/userHelper.js
@@ -1,11 +1,9 @@
 /* eslint-disable import/prefer-default-export */
-import Joi from 'joi';
 import jwt from 'jsonwebtoken';
 import { config } from 'dotenv';
 import bcrypt from 'bcrypt';
 import uuidv1 from 'uuidv1';
 import users from '../models/user';
-import articles from '../models/article';
 
 config();
 
@@ -38,53 +36,6 @@ const helper = {
     users.push(newUser);
   },
 
-  validateUser(user) {
-    const schema = {
-      firstName: Joi.string().min(3).required(),
-      lastName: Joi.string().min(3).required(),
-      email: Joi.string().email({ minDomainAtoms: 2 }),
-      password: Joi.string().regex(/^[a-zA-Z0-9]{3,30}$/),
-      gender: Joi.string().required(),
-      jobRole: Joi.string().min(3).required(),
-      department: Joi.string().required(),
-      address: Joi.string().required(),
-      isAdmin: Joi.boolean().required(),
-
-    };
-    return Joi.validate(user, schema);
-  },
-  validateArticle(article) {
-    const schema = {
-      authorId: Joi.string().required(),
-      authorName: Joi.string().required(),
-      title: Joi.string().min(3).required(),
-      article: Joi.string().min(10).required(),
-    };
-    return Joi.validate(article, schema);
-  },
-  validateComment(comment) {
-    const schema = {
-      authorName: Joi.string(),
-      authorId: Joi.string(),
-      comment: Joi.string().min(5).required(),
-    };
-    return Joi.validate(comment, schema);
-  },
-  findArticle(id) {
-    if (!id) {
-      // return;
-    }
-    const article = articles.find((art) => art.articleId === id);
-    return article;
-  },
-  checkAuthor(req) {
-    if (!req.payload) {
-      // return;
-    }
-    const author = this.getUser(req.payload.email);
-    const article = this.findArticle(req.params.id);
-    return author.id === article.authorId;
-  },
   securePassword(password) {
     if (!password) {
       // return;
